fix: pix.org errors causes nginx infinite loop
When accessing an unknown URL to pix.org (e.g. pix.org/fr/not-existing-page), Nginx try to access /404.html.
But the `rewrite` rules forces to 301 permanent redirect to /fr/404.html which doesn't exist and add /fr again.

Before finding a best solution, this is a small patch to avoid the Nginx infinite loop.